Move inline_svg helper to helpers directory
Middleman automatically loads helpers placed in this directory.
